+++ backend/routes/email.routes.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const emailServices = require("../services/email.services.js");
+
+router.post("/", emailServices.sendReport);
+
+module.exports = router;