+++ frontend/src/pages/ClientDetails/ClientDetailsPage.js
@@ -174,6 +174,7 @@ class ClientDetailsPage extends Component {
                                                 stylesMobile.openMapsLink
                                             }
                                         >
+                                            <i class="fa-solid fa-location-dot"></i>
                                             Voir sur Google Maps
                                         </a>
                                     </p>
@@ -239,6 +240,7 @@ class ClientDetailsPage extends Component {
                                                 stylesMobile.openMapsLink
                                             }
                                         >
+                                            <i class="fa-solid fa-location-dot"></i>
                                             Voir sur Google Maps
                                         </a>
                                     </p>
@@ -341,6 +343,7 @@ class ClientDetailsPage extends Component {
                                                 rel="noreferrer"
                                                 className={styles.openMapsLink}
                                             >
+                                                <i class="fa-solid fa-location-dot"></i>
                                                 Voir sur Google Maps
                                             </a>
                                         </p>
@@ -435,6 +438,7 @@ class ClientDetailsPage extends Component {
                                                 rel="noreferrer"
                                                 className={styles.openMapsLink}
                                             >
+                                                <i class="fa-solid fa-location-dot"></i>
                                                 Voir sur Google Maps
                                             </a>
                                         </p>